Moved string utilities into it's own header.
This is to enable using the string functions without including everything.
